+++ Guia 3/E3/DiscoHDD.cs
@@ -1,24 +1,10 @@
 namespace E3
 {
-    public class DiscoHDD:Componentes
+    public class DiscoHDD:Discos
     {
-        //conectores
-        private int precioDelComponente=0;
-        private string conectores;
-
-        public int PrecioDelComponente { get => precioDelComponente; }
-
-        public DiscoHDD(int precioDelComponente, string conectores)
+        public DiscoHDD(int precioDelComponente, string conectores):base(precioDelComponente,conectores)
         {
-            this.precioDelComponente = precioDelComponente;
-            this.conectores = conectores;
-        }
-
-        public int precio(){
-            return precioDelComponente;
-        }
-        public bool esCompatible(Motherboard motherboard){
-            return (conectores == motherboard.Conectores);
+  
         }
     }
 }
